requirements.txt
django
psycopg2-binary
djangorestframework
drf-nested-routers
django-filter
djangorestframework-simplejwt
django-cors-headers
rules
django-import-export